Refactor face detection into own module
